LGFS grad fee calculation
Javascript and service objects to call
Fee calculator API.
