+++ app/assets/javascripts/modules/external_users/claims/fee_calculator/FeeCalculator.js
@@ -1,0 +1,15 @@
+(function(exports, $) {
+  var Modules = exports.Modules.FeeCalculator || {};
+
+  // This init is called by the MOJ core js
+  Modules.init = function() {
+    if(this.UnitPrice){
+      this.UnitPrice.init();
+    }
+    if(this.GraduatedPrice){
+      this.GraduatedPrice.init();
+    }
+  };
+
+  exports.Modules.FeeCalculator = Modules;
+}(moj, jQuery));
