+++ app/views/external_users/claims/fixed_fees/litigators/_fields.html.haml
@@ -54,5 +54,5 @@
       .form-group
         %label.form-label-bold
           = t('.amount')
-        .fee-amount.total
+        .fee-net-amount.total
           = fee.amount || number_to_currency(0)
