+++ app/views/external_users/claims/graduated_fees/_fields.html.haml
@@ -1,53 +1,49 @@
 - if @claim.allows_graduated_fees?
   #graduated-fees
     = f.fields_for :graduated_fee do |ff|
-      - grad_fee = present(ff.object)
-      .form-section.graduated-fee-group.js-block.fx-do-init{data:{"type": "gradFees", autovat: @claim.apply_vat? ? "true" : "false", "block-type": "FeeBlockCalculator"}}
-        = ff.hidden_field :fee_type_id, value: @claim.case_type.graduated_fee_type.id
+      - fee = present(ff.object)
+      .form-section.fx-fee-group.graduated-fee-group.js-block.fx-do-init{data:{"type": "gradFees", autovat: @claim.apply_vat? ? "true" : "false"}}
+        = ff.hidden_field :fee_type_id, value: @claim.case_type.graduated_fee_type.id, class: 'js-fee-type'
 
         .form-section
           %a{:id => "graduated_fee.date"}
           = ff.gov_uk_date_field(:date, legend_text: "First day of hearing", legend_class: "govuk-legend", error_messages: gov_uk_date_field_error_messages(@error_presenter, "graduated_fee.date"))
 
+        .js-fee-calculator-effector
           = f.adp_text_field :actual_trial_length,
             label: t('.actual_trial_length'),
-            input_classes:'form-control form-control-1-4',
+            input_classes:'form-control form-control-1-4 js-fee-calculator-days',
             input_type: 'number',
             errors: @error_presenter
 
+        .js-fee-calculator-effector
           = ff.adp_text_field :quantity,
             label: t('.quantity'),
             hint_text: t('.quantity_hint'),
-            input_classes:'quantity fee-quantity form-control-1-4 js-fee-quantity js-fee-calculator-quantity"',
+            input_classes:'quantity fee-quantity form-control-1-4 js-fee-calculator-ppe',
             input_type: 'number',
-            value: grad_fee.quantity,
+            value: fee.quantity,
             errors: @error_presenter,
             error_key: "graduated_fee.quantity"
 
-          .js-fee-calculator-effectee
-            .calculated-fee
-            = ff.adp_text_field :rate,
-                label: t('.rate'),
-                input_classes:'rate fee-rate js-fee-calculator-rate form-input-denote__input form-control-1-4',
-                input_type: 'currency',
-                errors: @error_presenter,
-                input_readonly: grad_fee.rate_calculated?,
-                value: number_with_precision(grad_fee.rate, precision: 2)
-
           .js-fee-calculator-success
-            = ff.hidden_field :rate_calculated, value: grad_fee.rate_calculated?
+            = # ff.hidden_field :amount_calculated, value: fee.amount_calculated?
+
+          .js-fee-calculator-effectee
+            .calculated-grad-fee
+            = ff.adp_text_field :amount,
+              label: t('.amount'),
+              input_classes:"total fee-amount form-input-denote__input form-control-1-4",
+              input_type: "currency",
+              value: number_with_precision(fee.amount, precision: 2),
+              errors: @error_presenter,
+              error_key: "graduated_fee.amount"
 
           .help-wrapper.form-group{ style: 'display: none;' }
             %details
               %summary
                 %span.summary
-                  = t('.help_how_we_calculate_rate_title')
+                  = t('.help_how_we_calculate_amount_title')
               .panel.panel-border-narrow
                 %p
-                  = t('.help_how_we_calculate_rate_body')
-
-          .form-group
-            %label.form-label-bold
-              = t('.amount')
-            .fee-amount.total
-              = grad_fee.amount || number_to_currency(0)
+                  = t('.help_how_we_calculate_amount_body')
